refactor(type): splite some file
