+++ src/applepay/enum.ts
@@ -1,0 +1,3 @@
+export type FinalizeApplePayTransactionOptions = {
+  success: boolean;
+};
